+++ src/main/kotlin/astminer/parse/gumtree/GumtreeHandler.kt
@@ -3,18 +3,18 @@ package astminer.parse.gumtree
 import astminer.common.model.ParseResult
 import astminer.common.model.HandlerFactory
 import astminer.common.model.LanguageHandler
-import astminer.parse.gumtree.java.GumTreeJavaParser
-import astminer.parse.gumtree.java.GumTreeJavaFunctionSplitter
+import astminer.parse.gumtree.java.JDT.GumTreeJavaJDTParser
+import astminer.parse.gumtree.java.JDT.GumTreeJavaJDTFunctionSplitter
 import astminer.parse.gumtree.python.GumTreePythonFunctionSplitter
 import astminer.parse.gumtree.python.GumTreePythonParser
 import java.io.File
 
-object GumtreeJavaHandlerFactory : HandlerFactory {
-    override fun createHandler(file: File): LanguageHandler<GumTreeNode> = JavaGumtreeHandler(file)
+object GumtreeJavaJDTHandlerFactory : HandlerFactory {
+    override fun createHandler(file: File): LanguageHandler<GumTreeNode> = GumtreeJavaJDTHandler(file)
 
-    class JavaGumtreeHandler(file: File) : LanguageHandler<GumTreeNode>() {
-        override val splitter = GumTreeJavaFunctionSplitter()
-        override val parseResult: ParseResult<GumTreeNode> = GumTreeJavaParser().parseFile(file)
+    class GumtreeJavaJDTHandler(file: File) : LanguageHandler<GumTreeNode>() {
+        override val splitter = GumTreeJavaJDTFunctionSplitter()
+        override val parseResult: ParseResult<GumTreeNode> = GumTreeJavaJDTParser().parseFile(file)
     }
 }
