+++ src/test/kotlin/astminer/parse/gumtree/java/GumTreeJavaJDTParserTest.kt
@@ -1,15 +1,16 @@
 package astminer.parse.gumtree.java
 
 import astminer.common.getProjectFilesWithExtension
+import astminer.parse.gumtree.java.JDT.GumTreeJavaJDTParser
 import astminer.parseFiles
 import org.junit.Assert
 import org.junit.Test
 import java.io.*
 
-class GumTreeJavaParserTest {
+class GumTreeJavaJDTParserTest {
     @Test
     fun testNodeIsNotNull() {
-        val parser = GumTreeJavaParser()
+        val parser = GumTreeJavaJDTParser()
         val file = File("src/test/resources/examples/1.java")
 
         val node = parser.parseFile(file).root
@@ -18,7 +19,7 @@ class GumTreeJavaParserTest {
 
     @Test
     fun testProjectParsing() {
-        val parser = GumTreeJavaParser()
+        val parser = GumTreeJavaJDTParser()
         val projectRoot = File("src/test/resources/examples")
 
         val trees = parser.parseFiles(getProjectFilesWithExtension(projectRoot, "java"))
